Fix custom elements connect/disconnect spam
Fixes #219. When used with native custom elements distribution can cause an element to be disconnected/connected improperly. We avoid this by wrapping customElements.define and defining a class which manages connect/disconnect and calls through to super.* only if Shady DOM is not rendering the change. Specifically, render based disconnections are squelched and connections are only allowed if the element has not yet been connected or has been allowed to be disconnected.