Ensure ownerDocument is correct
Fixes #268. When a node in an inert document is appended to a node with a `shadowRoot` that has a `slot`, actual insertion is asynchronous. Because ownerDocument is not patched, this means actual ownerDocument is wrong when the custom elements polyfill tries to determine if the node should be upgraded. Since the ownerDocument is inert, the node is not upgraded.

To correct this, we always ensure ownerDocument is synchronously correct after an append via calling `adoptNode`.