Fixes #283
This could occur in rare cases when an element is created in an element constructor with a 'slot' child inside another shadowRoot.

This is due to the fact that a shadowRoot always renders the top of the distribution tree (https://github.com/webcomponents/shadydom/blob/master/src/attach-shadow.js#L118).

However, when rendering, nested shadowRoots in elements with slot children were only rendered if they had insertion points. They must also render if they are already pending rendering.