Don't allow the __cssBuild property to be renamed by Closure.
This was causing an incorrect interaction between webcomponents-lite.js
and apply-shim.js as an application dependency when property renaming is
enabled. Since they are not compiled together, and hence have different
rename maps, but both access the same property (since they both depend
on style-util.js and call getCssBuild), then it is never safe to rename
that property.