Add tracking of slot children and use to find distribution parent
This makes finding the top of the distribution tree more efficient. The distribution tree is defined as follows: if a shadowRoot's host has a slot child and its host's `getRootNode()` is a shadowRoot, this is its "distribution parent". When a shadowRoot renders, it always renders the highest shadowRoot in the distribution tree that is pending a render. This works because the "distribution parent" always renders all "distribution children".

Also adds a special skip to avoid any work on any queued render if `initialFlush` was called.