Preserve HTMLElement function name from jscompiler
This is a more comprehensive strategy to preserve the function name in the face of Closure compiler's optimizations.

Function.prototype.name will be inferred not only from the name give in the function expression/declaration itself, but also in the case of a function expression that's initializing a variable or property, it can be inferred to that variable or property name. Since closure compiler is free to remove the function name itself, but it isn't free to remove or rename a quoted property, we use that strategy here.