Adds `polyfillDefineLazy`
This is an optional optimization which allows defining an element with a function that returns a class. This function is called the first time the polyfill tries to upgrade an element of the given name. It allows the polyfill to defer any work needed to generate the element class or its callbacks until this work is actually needed.