Expose the installation code as a private global. (...)
This function should only be exposed when some kind of flag is set.
Unfortunately, this flag can't same type of flag that the polyfill normally
uses - properties set to a `customElements` global - because those flags may be
GC'd by the same bug that exposing this function is meant to work around.