Rewrite the deferred apply test: deterministic timing, async functions. (...)
Previously, the apply shim was loaded in a defer script and the test did not
necessarily wait for this script to load before running. (i.e. The test had a
very low (but non-zero) chance of failing if the apply shim loaded and ran
after the custom style in the page booted or the element was defined.) This
change updates the test to load the script at a known time (causing it to
fail), as well as use an async function.