Force connectedCallback / disconnected callback reactions to be run alternately. (...)
A flag `__CE_callbackStateConnected` is kept on all elements, indicating which
of `connectedCallback` or `disconnectedCallback` was last run, if either. When
another reaction is run on the element, if it is the same as the last reaction,
than another reaction of the opposite kind is run first so that they are always
called in alternating order.