Support shadow parts consuming custom properties
When a ::part rule consumes a custom property, the value of that property is
determined at each scope where that rule is _received_, not where the rule
was defined.

This means that, unless we have native custom properties (which already work
fine here with no changes), any time we find a part node that is receiving a
style that consumes a custom property, we need to switch that part's host to
per-instance styling, just as though it was consuming a custom property in its
own styles.

There is still a known issue where property values set in the exact same rule
that they are consumed are not supported. I included a skipped test for this and
will address in a follow up, since I need to do some more digging to find the
fix.