Add 'EnvironmentAPI' folder and add EventTarget API used by `watchFormdataTarget`.
In IE11 the user-facing API is not spec compliant and wrapping it in a way that
emulates using it as it was originally laid out is cumbersome. Particularly,
EventTarget is not a true participant in the prototype chain of objects that
extend it in spec IDL: `window.EventTarget` does not exist and functions that
are meant to be accessed from `window.EventTarget.prototype` are actually on
each subtype's prototype. So, to avoid capturing (e.g.)
`window.Node.prototype.addEventListener` and calling it on objects that aren't
Node instances, this folder should contain normalized versions of these
functions that expose a single `addEventListener` that is compatible with all
EventTarget subtypes.