Move part node discovery into Shady DOM.
Before this change, any mutation to a shimmed shadow root triggered a
querySelectorAll for "part" nodes that might be contained by a mutated shadow
root.

Now, we instead piggy-back on the existing walk that Shady DOM already does on
mutations (e.g. to look for slots), and look for "part" attributes there.