Add `finalSubmitListener`. (...)
This listener will be used to dispatch the 'formdata' event when the user calls
`stopPropagation` on a 'submit' event.

Normally, the 'formdata' event is dispatched when the 'submit' event reaches a
capturing listener on the shallow root of the tree containing the form; however,
the user can prevent the 'submit' event from reaching this node using
`stopPropagation` and `stopImmediatePropagation`. For
`stopImmediatePropagation`, the 'formdata' event can be dispatched immediately
inside a wrapper around `stopImmediatePropagation` because no other listeners
will be run. For `stopPropagation`, the 'formdata' event needs to be dispatched
*after* any other 'submit' events for the current phase that are attached to the
current target. `finalSubmitListener` is continually kept at the last position
in the list of listeners by `submitListenerAdded` and `submitListenerRemoved`
and will be used to dispatch the 'formdata' event if `stopPropagation` was
called for the event it sees.