Split `setSubmitEventPropagationStoppedCallback` into immediate and non-immediate cases and dispatch 'formdata' in the immediate case. (...)
Note that this means that the 'formdata' event is dispatched synchronously
during the call to `stopImmediatePropagation`. This means that any changes to
the form in the 'submit' event listener after this call will not be reflected in
the submitted data. This seems to imply that handling `stopPropagation` and
`stopImmediatePropagation` correctly would instead require all 'submit' event
listeners to be wrapped and these cases would be handled by the wrapper after
calling into the wrapped function.