Add `useNativeQuerySelector` setting. (...)
Normally, Shady DOM's wrappers for `querySelector` and `querySelectorAll` find
matches by manually walking the non-shadow-including descendants of the
reference element and calling `matches` on candidate elements. The new
`useNativeQuerySelector` setting causes them to instead call into the native
`querySelectorAll` and filter the results to only those within the same shadow
root as the reference element.

This setting trades the ability to find elements that are connected but
unassigned for the ability to use `:scope` pseudo-class in selectors:

- The default implementation is able to find nodes that are disconnected from
  the real tree because it chooses the set of candidate elements itself by
  walking. However, the meaning of `:scope` is changed for each native `matches`
  call during the walk because the reference element changes with each call,
  which makes `:scope` effectively unusable in selectors passed to the wrappers.

- The alternate implementation enabled by this setting is unable to find nodes
  that are disconnected from the real tree because the native `querySelectorAll`
  will only produce elements that are descendants of the reference element in
  the real tree. However, the meaning of `:scope` is unchanged because the
  reference element of the native call is the same as that of the wrapper call.