Fix closure compiler error
I'm trying to fix some bugs in the Closure compiler template type
checker. The fix has revealed a type error in shadydom, which this PR
attempts to fix.

With the fix, the following error occurs:

```
shadydom/src/patch-events.js:286:53: ERROR - [JSC_TYPE_MISMATCH] actual parameter 3 of Object.defineProperty does not match formal parameter
found   : ObjectPropertyDescriptor<Event>
required: ObjectPropertyDescriptor<(Event|{
  composed: ?,
  composedPath: function(this:Event): ?,
  relatedTarget: ?,
  stopImmediatePropagation: function(this:Event): undefined,
  stopPropagation: function(this:Event): undefined,
  target: ?
})>
  286|     /** @type {!ObjectPropertyDescriptor<!Event>} */ ({
                                                            ^^
  287|       /**
       ^^^^^^^^^
...
  296|       configurable: true,
       ^^^^^^^^^^^^^^^^^^^^^^^^^
  297|     })
       ^^^^^^

```

This error is correct because EventPatches is an object literal, which
can never be an Event because Event is a nominal type in Closure.

I think this change should be OK and will unblock Closure compiler
fixes.